create edit_hotkeys function
